move persistence to context?
