+++ src/contexts/use-persistence/persistence-contexts.ts
@@ -1,0 +1,7 @@
+import { createContext } from "react";
+
+import type { PersistenceContextType } from "./persistence-context";
+
+export const PersistenceContext = createContext<
+  PersistenceContextType | undefined
+>(undefined);
